add a readme and place config functions in separate file
